+++ changelog.rst
@@ -1,14 +1,19 @@
-1.0.0.1
+IN DEVELOPMENT
 -------
 2017-12-01
+
+- Changed: Assets folder to src folder
+- Added: .sass-lint.yaml for css linting
+- Changed: gulpfile when css gives a error don't quit but keep running
 
 - Changed: HTML template
 - Added: New gulp plugins
 - Changed: gulpfile.js
 - Changed: SASS files
 
+
 1.0.0.0
 -------
-2017-12-01
+2017-11-30
 
-- Inital release
+- Added: Inital release